backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
redis
celery
python-dotenv
openai
pgvector  # For RAG
langchain # For Agent Logic
langgraph # For Stateful Graphs
asyncpg
duckduckgo-search
psycopg2-binary